+++ packages/instantsearch.js/src/middlewares/__tests__/createInsightsMiddleware.ts
@@ -188,7 +188,7 @@ describe('insights', () => {
       expect(document.body).toMatchInlineSnapshot(`
         <body>
           <script
-            src="https://cdn.jsdelivr.net/npm/search-insights@2.15.0/dist/search-insights.min.js"
+            src="https://cdn.jsdelivr.net/npm/search-insights@2.17.2/dist/search-insights.min.js"
           />
         </body>
       `);
@@ -231,7 +231,7 @@ describe('insights', () => {
       expect(document.body).toMatchInlineSnapshot(`
         <body>
           <script
-            src="https://cdn.jsdelivr.net/npm/search-insights@2.15.0/dist/search-insights.min.js"
+            src="https://cdn.jsdelivr.net/npm/search-insights@2.17.2/dist/search-insights.min.js"
           />
         </body>
       `);
@@ -251,7 +251,7 @@ describe('insights', () => {
       expect(document.body).toMatchInlineSnapshot(`
         <body>
           <script
-            src="https://cdn.jsdelivr.net/npm/search-insights@2.15.0/dist/search-insights.min.js"
+            src="https://cdn.jsdelivr.net/npm/search-insights@2.17.2/dist/search-insights.min.js"
           />
         </body>
       `);