+++ packages/react-instantsearch-core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.13.8](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.13.7...react-instantsearch-core@7.13.8) (2024-11-14)
+
+
+### Bug Fixes
+
+* **deps:** update search-insights ([#6433](https://github.com/algolia/instantsearch/issues/6433)) ([05d93de](https://github.com/algolia/instantsearch/commit/05d93deb05d219f5d124175119d2e55707fb51c9)), closes [#6432](https://github.com/algolia/instantsearch/issues/6432)
+
+
+
+
+
 ## [7.13.7](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.13.6...react-instantsearch-core@7.13.7) (2024-11-05)
 
 **Note:** Version bump only for package react-instantsearch-core
@@ -109,7 +120,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Bug Fixes
 
 * **algoliasearch:** correctly retrieve headers for v5 ([#6268](https://github.com/algolia/instantsearch/issues/6268)) ([25c72cf](https://github.com/algolia/instantsearch/commit/25c72cf2f1957b72dc237b32a8a73a006753d5e1))
-* **deps:** update dependency search-insights to 2.15.0 ([#6273](https://github.com/algolia/instantsearch/issues/6273)) ([efeed39](https://github.com/algolia/instantsearch/commit/efeed39ae23a2ff244e40878a0a11eec5dd18211))
+* **deps:** update dependency search-insights to 2.17.2 ([#6273](https://github.com/algolia/instantsearch/issues/6273)) ([efeed39](https://github.com/algolia/instantsearch/commit/efeed39ae23a2ff244e40878a0a11eec5dd18211))